Isotropic material to separate module
